remove AccessListEntry interface and move its class to the datatypes module
