+++ ethereum/referencetests/src/main/java/org/hyperledger/besu/ethereum/referencetests/StateTestAccessListDeserializer.java
@@ -15,7 +15,7 @@
  */
 package org.hyperledger.besu.ethereum.referencetests;
 
-import org.hyperledger.besu.evm.AccessListEntry;
+import org.hyperledger.besu.datatypes.AccessListEntry;
 
 import java.io.IOException;
 import java.util.ArrayList;
